Add convert-rules CLI and move rules module to lib/
Add setup/convert-rules.mjs so Makefile can accept wildcard syntax
instead of requiring hand-written regex. Move rules.mjs and its tests
into setup/lib/ for better organization.
